Typo fix on misspelled + Int type fix

+++ src/main/scala/Utils/ClinksCalculator.scala
@@ -28,7 +28,7 @@ object ClinksCalculator:
     * @return n choose k
     */
   // TODO - Part 1 Step 1
-  def calculateCombination(n: Int, k: Int): BigInt = {
-    factorial(n)/factorial(k)/factorial(n-k)
+  def calculateCombination(n: Int, k: Int): Int = {
+    (factorial(n)/factorial(k)/factorial(n-k)).toInt
   }
 end ClinksCalculator
